Updated openstack/openstack Project: openstack/nova  a5943049c4fa90f8e0668d76361903b6ec0793f6
Prepare for unicode enums from Oslo

This tweaks our object version hash test a little bit. Oslo is getting
ready to coerce all Enum valid_values to unicode (which is what we should
have done in the first place). That will have no real runtime effect on
us, but our version hashes will all fail the minute we start running with
that code, just because of the way repr() signals unicode strings. This
just adds a hack to our test to pre-encode them to UTF-8 before we compare
them (on py2). On py3, everything is unicode already, so this doesn't
change anything there.

In the long(er) run, our Enum implementation will go away, and our test
that does the hash calculation will go away. So, this is mostly just a dance
to avoid us breaking the gate soon.

Change-Id: Ie20377e86134a5e8b68a93a08c26ed6f3025ae9d