Updated openstack/openstack Project: openstack/tempest  02753eea24910b391771d0ff210997a0c692d66b
Add unit tests for create/update/delete_agent

We have a plan to migrate service clients to tempest-lib from Tempest.
However, we are not doing enough tests for service clients now.
This patch adds unit tests for the methods create_agent, update_agent
and delete_agent.
In addition, this patch changes update_agent method because we will
stop truncating the top key and this change will be helpful for doing
that. Aslo this patch removes bytes casts because encode('utf-8')
returns bytes data, so current bytes calls are redundant.

Change-Id: I85cf43007c0654bb7747ff3518a19894b34271f8