Updated openstack/openstack Project: openstack/cinder  98e68a22c26db33f8471390bffa8dadde2d8fe4c
sqlalchemy exception kills FixedIntervalLoopingCall thread

Transient DB errors should not kill the thread for FixedIntervalLoopingCall.
This was noticed and fixed for the DBError case in Bug#1466991, but the
exception from sqlalchemy is not handled. Basically, self.model_disconnected
should be set to True, but the looping thread should carry on regardless of
the error.

Change-Id: I4bf04392269a2c33b77166b6c7408b38e1bc992d
Closes-Bug: #1478971