Updated openstack/openstack Project: openstack/cinder  d9c91126cd25458269583a88257d4494b5a8a55f
Use prefix for SolidFire template account

The SolidFire template account was not using the
account_prefix setting when the account was created.

This means that if a deployment is configured to use
the prefix, that calls to make the actual clone will fail
because the clone routine prepends the prefix from the
config settings on every lookup.

This patch just adds the same account adjustment that
we use in our regular account creation routines.

Change-Id: I4d3c22f96c9ca90192b88f1802997c8680b81483
Closes-Bug: #1479897