Updated openstack/openstack Project: openstack/nova  55b561d2d44c599416af106b66905e4f2457e079
libvirt:Add a driver API to inject an NMI

This patch adds a new driver API to inject an NMI to an instance.
This API is implemented by libvirt driver.

NMI (Non-maskable Interruption) is used to trigger a special function such as
the kernel crash dump mechanism.

DocImpact

Change-Id: Iaf63352139c466430d3304a87a66398133541fac
Implements: blueprint inject-nmi