Updated openstack/openstack Project: openstack/horizon  73341f6b9a35b0601d4e68756fc85fd7029c762e
Use form field errors on volume Extend form

Now when extending a volume to a new size which is less then the current
one, we will pop up an error warning message on the top of the form instead
of beside the new size field. now change to use field specified error class
for more accurate.

Closes-Bug: #1480001
Change-Id: I2ee1a562bf548e78d657a3c5bd3307eacf6a3ab1