Updated openstack/openstack Project: openstack/nova  d0d866bf41e374ce3c8ab057cc2093b79873e9d9
Allow NUMA based reporting for Monitors

This patchset allows monitors to report NUMA based metrics
It is a follow up to the work on creating a MonitorMetric Object.

Additionally, it defines two new fields to return the current
and maximum memory bw on a given NUMA node.

Co-Authored-By: Jay Pipes <jaypipes@gmail.com>

Change-Id: I6a0d589901851e392c75ab7ae754c454026e939f
Partially-Implements: blueprint memory-bw