Updated openstack/openstack Project: openstack-infra/puppet-zuul  89faadc064f26c6070bf3bafd4a3dee2c603257b
Log gerrit at INFO level

This is the normally expected configuration for Zuul logging.  This
reverts commit 5f780500c374221cb537389b6db1003042be30d5 from
the system-config repo.  DEBUG level was temporarily set to help
debug a Zuul issue in production.

Of course, future changes to OpenStack's Zuul logging should be
parameterized, but even so, the value in this change should be
the default.

Change-Id: I429346047d5901ca75afa04baddc7802605010ba