Updated openstack/openstack Project: openstack/ceilometer  6e969335d1cccfba51e0d0773f373a8546fe37cd
gnocchi: reduce the number of patch to gnocchi API

This refactor the processing of the samples to
more clearly separation the resource creation/update
and the measurements posting.

This refactor have permit to do only on PATCH of resource
per resource received into samples.

So, more samples are present in the received message more this code
is efficient.

Note, the workflow have changing, now resource are updated even
the posting of the measurements fail.

Change-Id: I01b758a13388da34636dc983a6c8c1550046d419