Updated openstack/openstack Project: openstack/horizon  f0e5cd501c1cee69d028feb8ee8267343923b2f1
Adding app.module.spec.js

This patch adds spec for app.module.js which has been missing since
it cannot be auto-collected to Jasmine test runner.

Added spec for the module's definition.  The run block is all about
configuring `horizon` object which is in old code and there is patch
moving them out off app.module.js since they are not really belong
to here, therefore specs for the code in the run block is not added.

Change-Id: I835eb6488ee5809ce874dbb7f5d34d06d0b5573d
Closes-Bug: #1480007