Updated openstack/openstack Project: openstack-infra/elastic-recheck  baed741f5d43cd68cdaf7056b206d3fe973eb95b
Add query for ceilometer meters faultstring 500

A newly added tempest test using ceilometer meters is failing at
roughly 3% of the time in the gate queue. (according to subunit2sql)
Ceilometer just stack traces without any useful debug information.
However, we can't use that stack trace here, because the ceilometer
api logs aren't indexed right now. This does the next best thing and
uses the one from the tempest failure traceback which actually contains
about as much info as the ceilo stack trace.

Right now it's showing 1 hit which is a failure in the gate queue, but
that's because logstash needs to catch up as this test was added
within the last few hours. (it's also not the failure I noticed this
in)

Change-Id: I3aaf62867ff995e4da998a5b89aff9efc96e107c
Related-Bug: #1480490