Updated openstack/openstack Project: openstack/api-site  7e91128a7b23e10b1b59bc07a6880e59f2816add
Delete an entry from "List security group rules" table

"Response parameters" table of "List security group rules" has an
unnecessary entry.  There are two "tenant_id" entries in the table.
The latter is the unnecessary one as "tenant_id" in a response
parameter is the ID of the tenant who owns the security group rule.

Change-Id: I98300d7af66f5085e7360b2fab3ee260440ac73d
Closes-Bug: #1477397