Updated openstack/openstack Project: openstack/keystone  b061972225090c8f8db6772da879606a0286d561
Reject create endpoint with invalid urls

The patch is not for general URL validation. It just implements
a check to make sure any substitutions in the URL will be valid
when building the catalog.

The bug also happen in v3. So we have to fix this bug in both
v2 and v3.

Closes-Bug: #1471034
Change-Id: I6728edaae26b140fc19afec6331674c57856985c