Updated openstack/openstack Project: openstack/glance  e45baa2915a1b520bf6fd190c1339abb14d67d71
glance metadef resource-type-associate fails in postgresql

The problem occurs only in postresql. The problem is in the default value
provided when creating a new resource-type. Instead of this line at line 182
in resource_type_association.py:
resource_type_dict = {'name': resource_type_name, 'protected': 0}
it should be 'protected': False for the proper code to be generated by
sqlalchemy.

Change-Id: I2115c64266a0def30ea466b1dc6d2d9db731645a
Closes-Bug: 1471048