Updated openstack/openstack Project: openstack/ironic  ba47899d59c16bcd1291b500a0b1b96a8584ac78
Save and re-raise exception

When an exception is raised, we do "some stuff" and re-raise that original
exception. This uses oslo_util's excutils.save_and_reraise_exception to
do this, to handle the case where "some stuff" causes the exception context
to be cleared (which would have resulted in None being attempted to be
re-raised).

Change-Id: I4b92915ae77fab9cde1787053152a497512401e2