Updated openstack/openstack Project: openstack/neutron  6fda25daa28743ad7d8e8156bb80f4e8b084764b
Move 1c844d1677f7 expand migration to appropriate branch

Commit cb60d0bb4e0cc0cba68f59fdf5f4e89d6ec52950 added an expand
migration. Howewer it indicated a contract migration as its
ancestor (2a16083502f3). As a result the migration was treated
by alembic as a contract one, even if it was defined in the
'expand' directory.

This patch 'rebases' this migration on top of the expand branch
HEAD. This might impact trunk chasers which have already executed
migration 1c844d1677f7 as a part of the contract branch.

Change-Id: I431aa684ccd029c6cff37a7666bb7ce1a7541f9c