Updated openstack/openstack Project: openstack/nova  de54fb4d97900ebd6d94318b26a1c9ecb27cbec7
cleanup NovaObjectDictCompat subclassing from pci_device

cleanup subclassing on NovaObjectDictCompat and fix subsequent
Unit tests and code associated with nova/objects/pci_device.py

Due to the exhaustive nature of changes, the cleanup is done
one object at a time.

This is the first patch of the series.
There shall be more patches to follow for other objects.

Related to blueprint liberty-objects

Change-Id: Idb764632e9a4f926bc1f159333de6a8366abcfab