Updated openstack/openstack Project: openstack/nova  771e8efc61bb3d625f2e0e839a3d14e378bd70cc
cells: document upgrade limitations/assumptions

We log a warning when using cells that it's experimental, and we have
several locations in the docs that say it's experimental, but we have
never documented that rolling upgrades are not supported and that cells
deployments must be upgraded in lockstep.  This adds that information to
the warning that's logged on startup and to the devref doc on upgrades
so people know that cells is still a special kitten that must be treated
differently.

Related-Bug: #1479181

Change-Id: Idd57f8daa308543a1733092f427e41b1a205b1d1