Updated openstack/openstack Project: openstack/barbican  1f27a56ce1e17fdd09c855c29cdc610e21d345c7
Add invalid property info to validation error message

It seems that the information that's being provided by our validation
errors is not sufficient for the users to make a sensible judgment about
what was the problem with their request to the server. This change
attempts to fix that by adding information about the specific property
that was problematic to the message that's given to the client.

Change-Id: I4bb72d2d2050bdd44aad2b89e71d50e4fe3d7f1f
Closes-bug: #1386251