Updated openstack/openstack Project: openstack/swift  708e9a5f05422cd9a43b094770b0ed957336e711
Make test_proxy work independent of env vars

test.unit.common.test_internal_client.TestSimpleClient.test_proxy
will fail if the environment has http_proxy set and does not have
no_proxy=127.0.0.1 set. This seems to be because urllib overrides
any proxy arg passed to the Request with the env http_proxy var,
unless the Request host is in no_proxy.

We don't need to test urllib, so this patch changes the test to
simply ensure that swift code does pass the correct proxy arg to
urllib2.urlopen. That avoids testers having to make sure their
env vars are compliant to successfully run unit tests. While it is
reasonable to think that 127.0.0.1 is in the no_proxy list, it
shouldn't be a requirement pass swift tests.

Change-Id: Iff91acdb76fabe7056b3e90e2756e27fe5105817