Updated openstack/openstack Project: openstack/cinder  b92e55e021eb3e04fb246165eca676b6eba14333
Fix get default quota values for subprojects

When getting the default quota value for a project, the DBQuotaDriver
is just checking if the default_quota_class CONF is set to True.
But this conf should not be considered when getting default quota
values for subprojects.

Previously, when we asked for the default quota values of a
subproject, the default quota values of the root project was returned.

Co-Authored-By: Vilobh Meshram <vilobhmm@yahoo-inc.com>

Change-Id: Ib88bafca915dfd387086c353c25b2af037e6d50f
Closes-bug: #1478651