Updated openstack/openstack Project: openstack/trove  387758bdcf1ab658f36496c89ef579d1f76a3416
Have devstack plugin install Trove client

Using the Trove devstack plugin was dependent on having the
python-troveclient also enabled.  This change modifies the plugin.sh
file to install the client correctly, thus removing the need of
separately enabling a plugin for it.

Variables were set up in the devstack settings file, which can be
overridden as desired.

A README.rst file was added with instructions on how to set up the
plugin in devstack.

Change-Id: If5a76360249827c19759a5c90dc00f70c59c32cc