Updated openstack/openstack Project: openstack/tempest  f565a0a84cf8cf4208270939292fecdc1aa67d46
Use relative path in data_files in setup.cfg

The absolute path causes failures when installing in a virtual
environment, not having root privileges.

This change solves the installation in a virtual environment,
but it does not fix yet the tempest init workflow when installing
into a virtual environment.

Partially-fixes bug #1480389

Change-Id: Iddefb8db77ffe35a84fb35b6f9e7e823de4ab45f