Updated openstack/openstack Project: openstack/heat  03ef858a71db05d4037369199dcf6b57188d175e
Convergence: Do create operation only if action is INIT

All resources that are new will have an INIT state. Instead
of having a complex strategy to decide whether the resource
should be created or updated, just check for the action
to see if it is in the INIT state or not. If it is not, then
always trigger the update workflow.

Also, this fixes a bug where we triggered a create for a
resource without a resource id that originally should've been
updated because it was in UPDATE_FAILED which was the unhandled
case.

Change-Id: I3f2318fecfe76592e8b54e9c09fdf1614197e83f