Updated openstack/openstack Project: openstack-infra/puppet-jeepyb  056388568343650fd51bc2be06836b0da1e6dca7
Fix parameters for jeepyb::openstackwatch

The swift_username, swift_password, swift_auth_url, and json_url
parameters all must be specified in order to work correctly, a default
value of empty string will cause brokenness. This patch removes the
default values for those parameters so that puppet will warn the user
that these values are required.

For the auth_version parameter, a default value of 1.0 is a safe
potential guess, so we change the default value from empty string to
'1.0'.

The feed parameter is no longer used as of
fdc8fd45bb26a8833662630c25b87b68d1e0ac12, so we remove it.

Change-Id: I0a10c0df80cb6a74e7bcc17d96e27e62373c61ff
Depends-On: Ie3c4f5184445b4653997c091279f21a3e9503b84