Updated openstack/openstack Project: openstack/oslo.vmware  596135d151a4df12114b878fb6dc1c5eb4297ca4
Allow http connection pool size to be configured

With multiple concurrent VM spawning, the service time from compute
service flow to vcenter to trigger soap requests is getting slower.
As VIM adopt requests module to access the vcenter which have default
http pool size set to 10. This need to be configured to be larger
in scale environment.

Change-Id: Ic9e9a87de378151a66187222231eaafd7a402ac5
Closes-Bug: #1479183