Updated openstack/openstack Project: openstack-infra/puppet-zuul  466040873b16ed4fa2cdf3ec567fbe04f52733c2
Adjust zuul's init.pp to check for and enable cgid instead of cgi.

When attempting to a2enmod cgi, Apache automatically selects cgid if a threaded
MPM is detected. The Puppet configuration then fails to converge when it is
unable to detect the expected cgi module in the future. Switching to explicit
cgid allows the configuration to converge without issues.

Change-Id: I71789d4a06c779541ea4abda5c467cd8c8729d0e