Updated openstack/openstack Project: openstack-infra/puppet-log_processor  3a57b1fe30ff42b9b0c65e1bc28dd96bd822a970
Add Gemfile and puppet 4 checks

In anticipation of puppet 4, start trying to deal with puppet 4 things
that can be helpfully predicted by puppet lint plugins. Also fix lint
errors caught by the puppet-lint-absolute_classname-check and
puppet-lint-empty_string-check gems.

This patch changes a scope.lookupvar function call to ruby instance
variable in the jenkins-log-client.default.erb template. This is safe
to do because the template is called in the log_processor::client class
and the variable in question is within that scope.

Change-Id: Ia7d6af8bc76a65e37f5dfd184e37855fe3b97046