Updated openstack/openstack Project: openstack/neutron  6c4669e4ded7f33ff59a0b9c830f1ceab3f8ebfa
Check that VXLAN is not in use in LB VXLAN check

The Linux bridge VXLAN supported check was only checking that the
test interface didn't exist instead of checking that both the interface
and the VXLAN didn't exist. This caused it to fail on startup if
a VXLAN interface existed under a different name using one of the
VXLANs that the agent tried to test support with.

This patch adds a check to ensure that the VXLAN ID isn't in use as well.

Closes-Bug: #1470579
Change-Id: I3a91ce54da86e319b7a4485dfae3fc99885383d4