Updated openstack/openstack Project: openstack-infra/puppet-zanata  df8f084120098fbc0d2dadff86e8f0fdeb79f6e4
Add second requirement for the ROOT.war deployment

Deploying puppet-zanata from scratch can sometimes fail because
wildfly has yet to be unpacked when puppet goes to copy the ROOT.war
into the deployment directory, which does not yet exist. Fix this
issue by adding a requirement for tar to run as well.

Change-Id: Id0229864b64ce39c70b559e1eb5d94822f1988ec
Co-Authored-By: Colleen Murphy <colleen@gazlene.net>