Updated openstack/openstack Project: openstack/nova  1d32e8d45d46c8579666edd804c4e77986258492
Base NovaObject on VersionedObject

This does the majority of the work towards getting NovaObject to be based on
oslo.versionedobjects' VersionedObject. It removes most of the meat of
NovaObject so that we use the equivalent methods in VersionedObject. It makes
minimal changes to the tests to show that the functionality is the same.
There are a few things still overridden in NovaObject that need less obvious
changes, which will come in subsequent patches, along with changes to drop
tests which are now redundant with o.vo coverage.

Related to blueprint use-oslo-objects

Change-Id: I208148319d4fef3a4c56ee1699b10bc3872f263f