Updated openstack/openstack Project: openstack/nova  a1e6abc125ed447ec2c97c1ac31b8b2d4ab4a1e9
Move to using ovo's remotable decorators

This removes Nova's remotable and remotable_classmethod decorators
and aliases the old location to the oslo.versionedobjects (ovo)
implementations.

This uncovered a couple other places where we were still passing a
context to a remotable method (in tests), which are fixed up here.

There is also a bug in ovo's remotable decorator that requires dict
compatibility. For this reason, this patch adds the Dict mixin to
a few objects (with appropriate comments), which can be removed when
we have an ovo release with that fix in it.

Related to blueprint use-oslo-objects

Change-Id: Id51fa3a55f65c8c075423f50a1240c997f9f2388