Updated openstack/openstack Project: openstack/nova  dfaf9b0a9dfa690c8171c7e93d477df9296549e6
Neutron: add in API method for updating VNIC index

The security group support for the VMware-NSXv Neutron plugin
requires that the Neutron port update the actual VNIC index of
the instance.

This patch adds the support for updating the VNIC index for the
Neutron port. This is done in two places:
1. When the instance is spawned
2. If and when a port is attached to a running instance

The Neutron support for the extension is:
https://review.openstack.org/#/c/144278

This is part of the blueprint vmware-nsxv-support

Change-Id: Ie990fa1141a7b2bdae1883572969adef072f4358