Updated openstack/openstack Project: openstack-infra/puppet-mediawiki  670beecdef3fb7de5695e26ba60e867724baaeb3
Fix parameter defaults

This patch aims to avoid defaulting parameters to the empty string and
to provide more reasonable default values where possible.

We set the default of $mediawiki::role to 'all' since that is what
Infra uses and is the most reasonable default between 'all', 'app', and
'image-scaler'.

We set the default of $mediawiki::site_hostname to $::fqdn instead of
empty string.

We set the default path of $mediawiki::mediawiki_location and
mediawiki::mediawiki_images_location to have base paths of
/srv/mediawiki since '/srv/mediawiki/w' is the hard-coded directory
path for the mediawiki repo.

We change the $mediawiki::ssl_cert_file and $mediawiki::ssl_key_file to
use the default snakeoil files so that a new user can easily set this
up without having to generate certs. This follows the pattern set by
the puppet-askbot and other modules.

We change the default values of the rest of the parameters from empty
string to undef and fix the logic involving those values.

Change-Id: Ic5271a952659896c75137b546c4b315c53c7511c