Updated openstack/openstack Project: openstack/tempest  4cb75f845f6b3de7c11ee9231ee4496071a9c8cf
Add plugin interface documentation

This commit starts the docs section on creating and using tempest
plugins. It describes how to create a plugin class, an entrypoint, and
a recommended dir structure. As well notes how to use tempest and
tempest-lib in creating a plugin. It'll likely be expanded in the
future as the plugin interface evolves.

Partially Implements bp external-plugin-interface

Change-Id: If2e0c47d0a719d869a3301ddbca07495c81cc498